backend-python/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
aspy.yaml==1.3.0
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.6.0
black==19.3b0
certifi==2019.11.28
cfgv==2.0.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
future==0.16.0
identify==1.4.10
idna==2.8
importlib-metadata==1.4.0
JPype1==0.7.1
konlpy==0.5.2
lxml==4.4.2
more-itertools==8.1.0
mysqlclient==1.4.6
nodeenv==1.3.4
numpy==1.18.1
oauthlib==3.1.0
pre-commit==1.16.1
pycodestyle==2.5.0
PySocks==1.7.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.9.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.23.4
tweepy==3.8.0
urllib3==1.25.7
virtualenv==16.7.9
zipp==2.0.0
